requirements.txt
pandas
scikit-learn
rdkit-pypi
joblib
matplotlib
seaborn
